--- movy/Features/Tab/ListScreenViewModel.swift
@@ -1,25 +1,0 @@
-//
-//  ListScreenViewModel.swift
-//  movy
-//
-//  Created by Maximiliano Ferraiuolo on 26/03/2025.
-//
-
-import SwiftUI
-
-class ListScreenViewModel: ObservableObject {
-    
-    // MARK: Public
-    
-    init() {
-      
-    }
-    
-    // MARK: Private
-      
-    // MARK: Static
-    
-    static func mock() -> ListScreenViewModel {
-        ListScreenViewModel()
-    }
-}
