+++ movy/Features/Tab/TabView.swift
@@ -13,30 +13,8 @@ struct MovyTabView: View {
   }
   
   var body: some View {
-    TabView(selection: $coordinator.currentTab) {
-        coordinator.homeCoordinator.build(page: .home)
-        .tag(TabsCoordinator.Tab.home)
-        .tabItem {
-          Label("Home", systemImage: "house")
-        }
-      
-        coordinator.searchCoordinator.build(page: .home)
-        .tag(TabsCoordinator.Tab.search)
-        .tabItem {
-          Label("Search", systemImage: "magnifyingglass")
-        }
-      
-      coordinator.listCoordinator.build(page: .home)
-        .tag(TabsCoordinator.Tab.list)
-        .tabItem {
-          Label("List", systemImage: "list.bullet")
-        }
-      
-      coordinator.profileCoordinator.build(page: .home)
-        .tag(TabsCoordinator.Tab.profile)
-        .tabItem {
-          Label("Profile", systemImage: "person.crop.circle")
-        }
-    }.tint(Color.blue)
+      TabView(selection: $coordinator.currentTab) {
+            coordinator.view(for: coordinator.currentTab)
+      }.tint(Color.blue)
   }
 }
