+++ movy/Features/Tab/Tabs/List/ViewModels/ListViewModel.swift
@@ -1,0 +1,31 @@
+//
+//  ListViewModel.swift
+//  movy
+//
+//  Created by Maximiliano Ferraiuolo on 30/03/2025.
+//
+
+
+import SwiftUI
+
+class ListViewModel: ObservableObject {
+
+    // MARK: Public
+
+    init( ) {
+      
+    }
+
+    func continueToNextStep() {
+        // TODO: Coordinator finish onboarding
+    }
+
+    // MARK: Private
+
+ 
+    // MARK: Static
+
+    static func mock() -> ListViewModel {
+        ListViewModel()
+    }
+}
