+++ movy/Features/Tab/Tabs/Profile/ViewModels/ProfileViewModel.swift
@@ -2,25 +2,26 @@
 //  ProfileViewModel.swift
 //  movy
 //
-//  Created by Maximiliano Ferraiuolo on 26/03/2025.
+//  Created by Maximiliano Ferraiuolo on 30/03/2025.
 //
 
 import SwiftUI
 
 class ProfileViewModel: ObservableObject {
-    
+
     // MARK: Public
-    
+
     init() {
-      
+
     }
-    
-    // MARK: Private
-      
+
+    func continueToNextStep() {
+        // TODO: Coordinator finish onboarding
+    }
+
     // MARK: Static
-    
+
     static func mock() -> ProfileViewModel {
         ProfileViewModel()
     }
 }
-
