+++ movy/Features/Tab/Tabs/Search/ViewModels/SearchViewModel.swift
@@ -2,24 +2,27 @@
 //  SearchViewModel.swift
 //  movy
 //
-//  Created by Maximiliano Ferraiuolo on 26/03/2025.
+//  Created by Maximiliano Ferraiuolo on 30/03/2025.
 //
 
 import SwiftUI
 
 class SearchViewModel: ObservableObject {
-    
+
     // MARK: Public
-    
+
     init() {
-      
+        
     }
-    
+
+    func continueToNextStep() {
+        // TODO: Coordinator finish onboarding
+    }
+
     // MARK: Private
-    
-    
+
     // MARK: Static
-    
+
     static func mock() -> SearchViewModel {
         SearchViewModel()
     }
